budget_service/requirements.txt
asgiref==3.8.1
confluent-kafka==2.10.0
Django==5.1.7
django-environ==0.12.0
djangorestframework==3.16.0
kafka-python==2.2.6
psycopg2-binary==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
